Backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cloudinary==1.44.0
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
gunicorn==23.0.0
packaging==25.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
